backend/api-server-ollama2rag/server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
langchain==0.3.7
langchain-community==0.3.7
qdrant-client==1.10.1
requests==2.32.3
tqdm==4.66.5
python-dotenv==1.0.1
PyPDF2==3.0.1
python-multipart==0.0.9
docxtpl==0.20.2
docx2pdf==0.1.8
python-docx==1.1.0
sqlmodel==0.0.14
aiosqlite==0.19.0